+++ packages/legacy/handlers/soap/src/index.ts
@@ -64,6 +64,8 @@ export default class SoapHandler implements MeshHandler {
           logger: this.logger,
           schemaHeaders: this.config.schemaHeaders,
           operationHeaders: this.config.operationHeaders,
+          soapHeaders: this.config.soapHeaders,
+          bodyAlias: this.config.bodyAlias,
         });
         const wsdlLocation = this.config.source;
         const wsdl = await readFileOrUrl<string>(wsdlLocation, {